split dashboard.js into actions + reducers + utils
