+++ frontend/src/metabase/dashboard/containers/DashboardApp.jsx
@@ -30,7 +30,7 @@ import {
 import { getDatabases, getMetadata } from "metabase/selectors/metadata";
 import { getUserIsAdmin } from "metabase/selectors/user";
 
-import * as dashboardActions from "../dashboard";
+import * as dashboardActions from "../actions";
 import { parseHashOptions } from "metabase/lib/browser";
 import * as Urls from "metabase/lib/urls";
