+++ frontend/src/metabase/dashboard/hoc/DashboardData.jsx
@@ -13,7 +13,7 @@ import {
   getParameterValues,
 } from "metabase/dashboard/selectors";
 
-import * as dashboardActions from "metabase/dashboard/dashboard";
+import * as dashboardActions from "metabase/dashboard/actions";
 
 import type { Dashboard } from "metabase-types/types/Dashboard";
 import type { Parameter } from "metabase-types/types/Parameter";
