+++ frontend/src/metabase/reducers-public.js
@@ -2,7 +2,7 @@
 
 import commonReducers from "./reducers-common";
 
-import dashboard from "metabase/dashboard/dashboard";
+import dashboard from "metabase/dashboard/reducers";
 
 export default {
   ...commonReducers,
